Move test cases for decoding event data and key to 'src/onchain_events/index.spec.ts'
